Refactor timestamp metrics
